cryptlib: Fix memory leak and bounds
Program was crashing due to memory issues.

Upon investigation these crashes were determined to be caused by going out of bounds due to not copying the null-terminator.

A temporary solution has been implemented replacing strncpy with strcpy. If a more complicated method is needed in the future it can be replaced at such time. There was also a memory leak related to this where free was being called too much and as part of investigation was called too little. This has been now changed to be called just the right amount of times.

Signed-off-by: David Piper <dbpiper@gmail.com>