bytes: Create bytes struct, refactor to use it.
Code was getting ugly passing around unsigned char * and int all the time.

Merging these two into a bytes struct helps to improve readability, usability, and removes bulk from the code. Refactoring to incorporate this change made a lot of old code in dataconvert obsolete as I decided to simply have init code in the bytes module to init it from an ascii string or hex string. This could even be improved in the future with Unicode support, but this will likely not happen for some time.

With the bytes struct, I was able to cleanly test bytetok from dataconvert. Score.c has been updated with the fixes discovered during the process of changing and fixing bytetok through TDD. As bytetok was by far the most complex and difficult to test function in dataconvert, it should not be much longer before it is fully tested.

Signed-off-by: David Piper <dbpiper@gmail.com>