cryptlib: Reimplement Challenge 5
Adds proper support for Challenge 5

Until now, even before the refactor, Challenge 5 was handled in a way that was a hack at best and broken at worst. This has been fixed by proper support, and was partially the motivation for the refactor in the first place.

As a result, the solution is much more elegant and easier to read. This was accomplished by having a function in datamanip to handle repeating xor and functions in dataconvert to handle convert each byte to two zero-padded hex digits.

Signed-off-by: David Piper <dbpiper@gmail.com>